+++ challenges/linked-list/linked_list.py
@@ -2,38 +2,86 @@ from node import Node
 
 
 class LinkedList:
-    """This will be a link list class """
+    """ class for new linked list """
     def __init__(self, iter=[]):
-        self._current = None
         self.head = None
-        self._size = 0
+        self._len = 0
 
-# l = LinkedList([1, 2, 3, 4])
+        for item in iter:
+            self.head = Node(item, self.head)
+            self._len += 1
 
+    def __len__(self):
+        """ return len of the current object """
+        return self._len
 
-for item in reversed(iter):
-    self.insert(item)
-# l.head => 1 -> 2 -> 3 -> 4 ->
+    def __str__(self):
+        """ return all items from the LL """
+        lis = []
+        current = self.head
+        for _ in range(self._len+1):
+            lis.append(current)
+            current = current._next
+        return str(lis)
 
+    def insert(self, val):
+        """ add item to LL """
+        self.head = Node(val, self.head)
+        self._len += 1
 
-def __repr__(self):
-    # assuming head will have a val (- need to handle the case of None)
-    return '<head> => {}' .format(self.head.val)
+    def find(self, val):
+        """ search for element and return TRUE or False"""
+        x = self.head
+        while x is not None:
+            if x == val:
+                return True
+            return False
 
+    def append(self, value):
+        """ append vlue at the end of the lsit """
+        current = self.head
+        while current._next:
+            current = current._next
+        current._next = Node(value)
 
-def __len__(self):
-    return self.size
+    def insert_before(self, val, newVal):
+        """ insert new node before current """
+        if self._size == 0:
+            return 'Linked List is Empty'
+        previous_node = self.head
+        self.head = Node(newVal, previous_node)
+        return self
+        if previous_node._next:
+            x = previous_node._next
+            while x:
+                if x.val == val:
+                    print('x', x)
+                    new_insert = Node(newVal, x)
+                    previous_node._next = new_insert
+                    self._size += 1
+                    print('length => {}'.format(self.__len__()))
+                    return self
+                previous_node = x
+                x = x._next
+        return 'did not insert'
 
-
-def __str__(self):
-
-
-def find(self, val):
-
-
-def insert(self, val):
-    self.head = Node(val, self.head)
-    self._size += 1
-
-
-# still need to handle error handling
+    def insert_after(self, val, newVal):
+        """ Insert new node after current """
+        if self._size == 0:
+            return 'Linked List is Empty'
+        previous_node = self.head
+        self.head = Node(newVal, previous_node)
+        return self
+        if previous_node._next:
+            x = previous_node._next
+            while x:
+                if x.val == val:
+                    print('x', x)
+                    new_insert = Node(newVal, x)
+                    previous_node._next = new_insert
+                    self._size += 1
+                    print('length => {}'.format(self._len__()))
+                    return self
+                previous_node = x
+                x = x._next
+        return 'did not insert'
